+++ src/Core/Responses/RelationshipResponse.php
@@ -22,6 +22,9 @@ namespace LaravelJsonApi\Core\Responses;
 use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Http\Request;
 use LaravelJsonApi\Contracts\Pagination\Page;
+use LaravelJsonApi\Core\Responses\Concerns\HasEncodingParameters;
+use LaravelJsonApi\Core\Responses\Concerns\HasRelationshipMeta;
+use LaravelJsonApi\Core\Responses\Concerns\IsResponsable;
 use LaravelJsonApi\Core\Responses\Internal\PaginatedIdentifierResponse;
 use LaravelJsonApi\Core\Responses\Internal\ResourceCollectionResponse;
 use LaravelJsonApi\Core\Responses\Internal\ResourceIdentifierCollectionResponse;
@@ -32,8 +35,9 @@ use function is_null;
 class RelationshipResponse implements Responsable
 {
 
-    use Concerns\HasRelationshipMeta;
-    use Concerns\IsResponsable;
+    use HasEncodingParameters;
+    use HasRelationshipMeta;
+    use IsResponsable;
 
     /**
      * @var object
@@ -91,6 +95,8 @@ class RelationshipResponse implements Responsable
             ->withMeta($this->meta)
             ->withLinks($this->links)
             ->withEncodeOptions($this->encodeOptions)
+            ->withIncludePaths($this->includePaths)
+            ->withSparseFieldSets($this->fieldSets)
             ->withHeaders($this->headers());
     }